trimmomatic exceeded soft mem limit
Ali, how long would you expect this to take and how much memory should we give it? it seemed like it was taking a long time and it exceeded the soft mem limit with 4 gb. we should put a note in this document about how  much to give the job.